The Big Shrinking Begins.
We're switching to int hashMultiplier values because long really is *that much slower* on GWT. As long as we have imul, we're fine.

This only gets the Object-keyed and Identity-based maps and sets. More are coming soon!

+++ src/main/java/com/github/tommyettinger/ds/IdentityObjectMap.java
@@ -144,7 +144,7 @@ public class IdentityObjectMap<K, V> extends ObjectObjectMap<K, V> {
 	 * @return the current hash multiplier
 	 */
 	@Override
-	public long getHashMultiplier () {
+	public int getHashMultiplier () {
 		return hashMultiplier;
 	}
 
@@ -155,8 +155,8 @@ public class IdentityObjectMap<K, V> extends ObjectObjectMap<K, V> {
 	 * @param hashMultiplier will not be used, but will be treated as odd and stored in case some other code needs it
 	 */
 	@Override
-	public void setHashMultiplier (long hashMultiplier) {
-		this.hashMultiplier = hashMultiplier | 1L;
+	public void setHashMultiplier (int hashMultiplier) {
+		this.hashMultiplier = hashMultiplier | 1;
 	}
 
 	/**
