+++ src/main/java/com/github/tommyettinger/ds/IdentityOrderedSet.java
@@ -94,7 +94,7 @@ public class IdentityOrderedSet<T> extends ObjectOrderedSet<T> {
 	 * @return the current hash multiplier
 	 */
 	@Override
-	public long getHashMultiplier () {
+	public int getHashMultiplier () {
 		return hashMultiplier;
 	}
 
@@ -105,8 +105,8 @@ public class IdentityOrderedSet<T> extends ObjectOrderedSet<T> {
 	 * @param hashMultiplier will not be used, but will be treated as odd and stored in case some other code needs it
 	 */
 	@Override
-	public void setHashMultiplier (long hashMultiplier) {
-		this.hashMultiplier = hashMultiplier | 1L;
+	public void setHashMultiplier (int hashMultiplier) {
+		this.hashMultiplier = hashMultiplier | 1;
 	}
 
 	public static <T> IdentityOrderedSet<T> with (T item) {
