+++ src/main/java/com/github/tommyettinger/ds/IdentitySet.java
@@ -89,7 +89,7 @@ public class IdentitySet<T> extends ObjectSet<T> {
 	 * @return the current hash multiplier
 	 */
 	@Override
-	public long getHashMultiplier () {
+	public int getHashMultiplier () {
 		return hashMultiplier;
 	}
 
@@ -100,8 +100,8 @@ public class IdentitySet<T> extends ObjectSet<T> {
 	 * @param hashMultiplier will not be used, but will be treated as odd and stored in case some other code needs it
 	 */
 	@Override
-	public void setHashMultiplier (long hashMultiplier) {
-		this.hashMultiplier = hashMultiplier | 1L;
+	public void setHashMultiplier (int hashMultiplier) {
+		this.hashMultiplier = hashMultiplier | 1;
 	}
 
 	public static <T> IdentitySet<T> with (T item) {
